🧹 extract os filesystem connection into submodule
